WIP Implement skeletal animation: move Mesh classes to separate header.
