+++ src/app/gltf_viewer/gltf_viewer/Camera.h
@@ -1,6 +1,9 @@
 #pragma once
 
 
+#include "Logging.h"
+#include "Polar.h"
+
 #include <graphics/AppInterface.h>
 #include <graphics/CameraUtilities.h>
