+++ src/app/gltf_viewer/gltf_viewer/Scene.cpp
@@ -10,41 +10,9 @@ namespace gltfviewer {
 //
 // Programs
 //
-void Scene::initializePrograms()
-{
-    {
-        auto naive = std::make_shared<graphics::Program>(
-            graphics::makeLinkedProgram({
-                {GL_VERTEX_SHADER,   gltfviewer::gNaiveVertexShader},
-                {GL_FRAGMENT_SHADER, gltfviewer::gNaiveFragmentShader},
-            }));
-        shaderPrograms.push_back(std::move(naive));
-    }
-
-    {
-        auto phong = std::make_shared<graphics::Program>(
-            graphics::makeLinkedProgram({
-                {GL_VERTEX_SHADER,   gltfviewer::gPhongVertexShader},
-                {GL_FRAGMENT_SHADER, gltfviewer::gPhongFragmentShader},
-            }));
-        setUniform(*phong, "u_light.position_world", math::Vec<4, GLfloat>{-100.f, 100.f, 1000.f, 1.f}); 
-        setUniform(*phong, "u_light.color", math::hdr::gWhite<GLfloat>);
-        setUniformInt(*phong, "u_light.specularExponent", 100);
-        // Ideally, I suspect the sum should be 1
-        setUniformFloat(*phong, "u_light.diffuse", 0.3f);
-        setUniformFloat(*phong, "u_light.specular", 0.35f);
-        setUniformFloat(*phong, "u_light.ambient", 0.45f);
-        shaderPrograms.push_back(std::move(phong));
-    }
-}
-
-
 void Scene::setProjection(const math::Matrix<4, 4, float> & aProjectionTransform)
 {
-    for (auto & program : shaderPrograms)
-    {
-        setUniform(*program, "u_projection", aProjectionTransform); 
-    }
+    renderer.setProjectionTransformation(aProjectionTransform);
     debugDrawer.setProjectionTransformation(aProjectionTransform);
 }
 
@@ -77,7 +45,7 @@ Scene::getBoundingBox(arte::Const_Owned<arte::gltf::Node> aNode,
     std::optional<math::Box<GLfloat>> result;
     if(aNode->mesh)
     {
-        result = indexToMeshes.at(*aNode->mesh).mesh.boundingBox * modelTransform;
+        result = indexToMesh.at(*aNode->mesh).mesh.boundingBox * modelTransform;
     }
     
     for (auto node : aNode.iterate(&arte::gltf::Node::children))
